refactor(dbt): separate out `DbtCliInvocation` to `dbt_cli_invocation.py`
