+++ python_modules/libraries/dagster-dbt/dagster_dbt/__init__.py
@@ -18,11 +18,9 @@ from .cloud import (
     dbt_cloud_run_op as dbt_cloud_run_op,
     load_assets_from_dbt_cloud_job as load_assets_from_dbt_cloud_job,
 )
-from .core.resource import (
-    DbtCliEventMessage as DbtCliEventMessage,
-    DbtCliInvocation as DbtCliInvocation,
-    DbtCliResource as DbtCliResource,
-)
+from .core.dbt_cli_event import DbtCliEventMessage as DbtCliEventMessage
+from .core.dbt_cli_invocation import DbtCliInvocation as DbtCliInvocation
+from .core.resource import DbtCliResource as DbtCliResource
 from .dagster_dbt_translator import (
     DagsterDbtTranslator as DagsterDbtTranslator,
     DagsterDbtTranslatorSettings as DagsterDbtTranslatorSettings,
